Move handler class to private class in exposition.py
